🚚 api: move level seven info route to evaluation context
