+++ api/lib/application/users/index.js
@@ -525,31 +525,6 @@ const register = async function (server) {
     },
     {
       method: 'PATCH',
-      path: '/api/users/{id}/user-has-seen-level-seven-info',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkRequestedUserIsAuthenticatedUser,
-            assign: 'requestedUserIsAuthenticatedUser',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.userId,
-          }),
-        },
-        handler: userController.rememberUserHasSeenLevelSevenInfo,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Sauvegarde le fait que l'utilisateur ait vu le message d'information d'ouverture du niveau 7" +
-            '- L’id demandé doit correspondre à celui de l’utilisateur authentifié',
-          "- Le contenu de la requête n'est pas pris en compte.",
-        ],
-        tags: ['api', 'user'],
-      },
-    },
-    {
-      method: 'PATCH',
       path: '/api/users/{id}/has-seen-challenge-tooltip/{challengeType}',
       config: {
         pre: [
